refactor(build): remove core component
