+++ components/CMakeLists.txt
@@ -1,5 +1,4 @@
 set(PROJECT_COMPONENTS
-    core
     settings_core
     settings_ui
     connection_tester
