--- components/core/CMakeLists.txt
@@ -1,7 +1,0 @@
-idf_component_register(
-    SRCS
-        "src/ringbuf.cpp"
-        "src/settings.cpp"
-    INCLUDE_DIRS
-        "include"
-)
